test: align test/fields with TEST_STYLE (placement, dedup, tighter asserts)
Relocate tests so test/ mirrors src/ 1:1, one home per source module:
- Model.field(...) tests -> test/framework/model/test_spec.py (Cycle 7)
- run_build auto-synthesis tests -> test/framework/model/test_spec.py (Cycle 8)
- configurations field-schema tests -> test/framework/solver/test_configurations.py
- rename test_auto_synthesis.py -> test_synthesis.py (synthesis-only)

Dedup: shared staged_hot_room fixture in new test/fields/conftest.py.
Tighten asserts: FrozenInstanceError / ValidationError, add match= on
empty-allowed_bcs, unsupported-value-type, and empty-union raises.
Hoist in-function imports to module top.

Pure test refactor: 147 tests conserved (relocated, not lost/duplicated);
pre-commit clean on all touched files.
