+++ test/fields/test_loader.py
@@ -12,18 +12,15 @@ real OpenFOAM dictionaries from disk, not encoded strings).
 
 from __future__ import annotations
 
-import shutil
 from pathlib import Path
 
 import pytest
+from pydantic import BaseModel
 
 from neofoam.fields.bc import FixedValueBC, GenericBC, NoSlipBC, ZeroGradientBC
 from neofoam.fields.loader import load_fields, save_fields
 from neofoam.fields.value_types import Scalar, Vector
 from neofoam.framework.model.spec import Model
-
-
-FIXTURE = Path(__file__).parent / "cases" / "hotRoom" / "0.orig"
 
 
 class _SyntheticSolver:
@@ -104,47 +101,31 @@ def _hot_room_solver() -> _SyntheticSolver:
     return _SyntheticSolver("hotRoom", [pimple, boussinesq, turbulence])
 
 
-# -- helpers -----------------------------------------------------------
-
-
-def _stage(tmp_path: Path) -> Path:
-    """Copy ``0.orig/*`` into ``tmp_path/0/`` for a writable case."""
-    case = tmp_path / "case"
-    (case / "0").mkdir(parents=True)
-    for f in FIXTURE.iterdir():
-        shutil.copy(f, case / "0" / f.name)
-    return case
-
-
 # -- tests -------------------------------------------------------------
 
 
-def test_load_fields_covers_every_declared_name(tmp_path: Path) -> None:
-    case = _stage(tmp_path)
-    fields = load_fields(case, solver=_hot_room_solver())
+def test_load_fields_covers_every_declared_name(staged_hot_room: Path) -> None:
+    fields = load_fields(staged_hot_room, solver=_hot_room_solver())
     assert set(fields) == {"U", "p", "p_rgh", "T", "alphat", "k", "epsilon", "nut"}
 
 
-def test_loaded_dimensions_match_disk(tmp_path: Path) -> None:
-    case = _stage(tmp_path)
-    fields = load_fields(case, solver=_hot_room_solver())
+def test_loaded_dimensions_match_disk(staged_hot_room: Path) -> None:
+    fields = load_fields(staged_hot_room, solver=_hot_room_solver())
     assert fields["U"].dimensions == [0, 1, -1, 0, 0, 0, 0]
     assert fields["T"].dimensions == [0, 0, 0, 1, 0, 0, 0]
     assert fields["p_rgh"].dimensions == [0, 2, -2, 0, 0, 0, 0]
 
 
-def test_typed_arm_dispatch(tmp_path: Path) -> None:
+def test_typed_arm_dispatch(staged_hot_room: Path) -> None:
     """U has only noSlip BCs in the fixture — the typed arm wins."""
-    case = _stage(tmp_path)
-    fields = load_fields(case, solver=_hot_room_solver())
+    fields = load_fields(staged_hot_room, solver=_hot_room_solver())
     for patch, bc in fields["U"].boundaryField.items():
         assert isinstance(bc, NoSlipBC), (patch, bc)
 
 
-def test_generic_fallback_keeps_unknown_bcs_intact(tmp_path: Path) -> None:
+def test_generic_fallback_keeps_unknown_bcs_intact(staged_hot_room: Path) -> None:
     """p_rgh uses fixedFluxPressure with extras; GenericBC preserves them."""
-    case = _stage(tmp_path)
-    fields = load_fields(case, solver=_hot_room_solver())
+    fields = load_fields(staged_hot_room, solver=_hot_room_solver())
     for patch, bc in fields["p_rgh"].boundaryField.items():
         assert isinstance(bc, GenericBC), (patch, bc)
         dumped = bc.model_dump()
@@ -152,40 +133,35 @@ def test_generic_fallback_keeps_unknown_bcs_intact(tmp_path: Path) -> None:
         assert dumped["rho"] == "rhok"
 
 
-def test_round_trip_preserves_payload(tmp_path: Path) -> None:
+def test_round_trip_preserves_payload(staged_hot_room: Path) -> None:
     """load_fields → save_fields → reload should produce the same dump."""
-    case = _stage(tmp_path)
     solver = _hot_room_solver()
 
-    before = load_fields(case, solver=solver)
-    save_fields(before, case)
-    after = load_fields(case, solver=solver)
+    before = load_fields(staged_hot_room, solver=solver)
+    save_fields(before, staged_hot_room)
+    after = load_fields(staged_hot_room, solver=solver)
 
     for name in before:
         assert before[name].model_dump() == after[name].model_dump(), name
 
 
-def test_mutation_before_save_is_persisted(tmp_path: Path) -> None:
-    case = _stage(tmp_path)
+def test_mutation_before_save_is_persisted(staged_hot_room: Path) -> None:
     solver = _hot_room_solver()
 
-    fields = load_fields(case, solver=solver)
+    fields = load_fields(staged_hot_room, solver=solver)
     # Flip the floor BC on T from fixedValue → zeroGradient; reload and assert.
     fields["T"].boundaryField["floor"] = ZeroGradientBC()
-    save_fields({"T": fields["T"]}, case)
+    save_fields({"T": fields["T"]}, staged_hot_room)
 
-    reloaded = load_fields(case, solver=solver)
+    reloaded = load_fields(staged_hot_room, solver=solver)
     assert isinstance(reloaded["T"].boundaryField["floor"], ZeroGradientBC)
     # The other patches survive untouched.
     assert isinstance(reloaded["T"].boundaryField["ceiling"], FixedValueBC)
 
 
-def test_save_fields_rejects_unbound_instance(tmp_path: Path) -> None:
-    case = _stage(tmp_path)
-    from pydantic import BaseModel
-
+def test_save_fields_rejects_unbound_instance(staged_hot_room: Path) -> None:
     class _Unbound(BaseModel):
         pass
 
     with pytest.raises(ValueError, match="no io_config"):
-        save_fields({"x": _Unbound()}, case)  # type: ignore[dict-item]
+        save_fields({"x": _Unbound()}, staged_hot_room)  # type: ignore[dict-item]
